Fix subsetting error for Legendre polynomials
The default behavior of `leg` names the columns of the coefficient matrix (leg0, leg1, etc.), which are used to construct the design matrix in the call to `vs`. When the user requests a first order polynomial without the intercept term, the default behavior of the subset operator when a single column is requested from a matrix is to return a vector. Consequently, the column name is lost during the subset and not replaced after the call to `as.matrix`. Removing the call to `as.matrix` and adding the `drop = FALSE` argument to the subset operator preserves the matrix structure and the column name even when keeping a single column from the original coefficient matrix.